[FFM-9071] - Rewrite flag/targetsegment cache
What
This is an update to the SDK's usage of the Platform Cache. Instead of loading all flag/segment data into 2 cache keys this change spreads the data across multiple keys.

Why
For customers with a lot of flags we're seeing Platform Cache limits of 100K being reached very quickly, which is causing an exception to be thrown and the SDK to fail in these cases.

Testing
Manual + unit